8.1 Big update
Added text during the transition.

Added a wallet.

Added a change in the amount of money in the game menu and in the menu for buying a place to shoot.

Added a text change on the level slider.

Added the ability to buy additional shooting spots.

The working principle of the buttons has been changed.